requirements.txt
aniso8601==2.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-RESTful==0.3.6
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.1
opencv-python==3.4.0.12
pdfminer.six==20170720
Pillow==5.0.0
pkg-resources==0.0.0
pycryptodome==3.4.11
pytesseract==0.2.0
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
